Move data-driven test infrastructure to something that is shippable
