+++ neuro_san/test/evaluators/gist_agent_evaluator.py
@@ -16,8 +16,7 @@ from neuro_san.client.agent_session_factory import AgentSessionFactory
 from neuro_san.client.streaming_input_processor import StreamingInputProcessor
 from neuro_san.interfaces.agent_session import AgentSession
 from neuro_san.message_processing.basic_message_processor import BasicMessageProcessor
-
-from tests.framework.evaluators.abstract_agent_evaluator import AbstractAgentEvaluator
+from neuro_san.test.evaluators.abstract_agent_evaluator import AbstractAgentEvaluator
 
 
 class GistAgentEvaluator(AbstractAgentEvaluator):
