+++ neuro_san/test/unittest/dynamic_hocon_unit_tests.py
@@ -14,9 +14,8 @@ from typing import List
 from unittest import TestCase
 
 from neuro_san.internals.utils.file_of_class import FileOfClass
-
-from tests.framework.driver.data_driven_agent_test_driver import DataDrivenAgentTestDriver
-from tests.framework.unittest.unit_test_assert_forwarder import UnitTestAssertForwarder
+from neuro_san.test.driver.data_driven_agent_test_driver import DataDrivenAgentTestDriver
+from neuro_san.test.unittest.unit_test_assert_forwarder import UnitTestAssertForwarder
 
 
 class DynamicHoconUnitTests:
