+++ tests/neuro_san/zzz_hocons/test_unit_test_hocons.py
@@ -13,7 +13,7 @@ from unittest import TestCase
 
 from parameterized import parameterized
 
-from tests.framework.unittest.dynamic_hocon_unit_tests import DynamicHoconUnitTests
+from neuro_san.test.unittest.dynamic_hocon_unit_tests import DynamicHoconUnitTests
 
 
 class TestUnitTestHocons(TestCase):
